move db port to ports module
